requirements.txt
# Core bot framework
aiogram==3.18.0
asyncio-mqtt==0.16.2

# Database
tortoise-orm==0.24.2
aiosqlite==0.21.0

# Google APIs
gspread==6.2.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
oauth2client==4.1.3

# OpenAI
openai==1.66.5

# Scheduling
APScheduler==3.11.0

# HTTP and utilities
aiohttp==3.11.14
aiofiles==24.1.0
requests==2.32.3
python-dateutil==2.9.0

# Logging and monitoring
structlog==24.4.0

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Environment management
python-dotenv==1.0.1